changes, not tested
