+++ traffic-sniffer/.gitignore
@@ -8,4 +8,5 @@ dist
 sharkio-0.1.0.tgz
 *.tgz
 .idea
-config/sniffers-setup.json
+
+sniffers-setup.json
