+++ traffic-sniffer/lib/request-metadata/setup-config/config-loader-interface.ts
@@ -1,0 +1,17 @@
+import { SnifferConfig } from "../sniffer/sniffer";
+import { SnifferConfigSetup } from "./file-config";
+
+
+export interface ConfigLoader {
+  configData: SnifferConfigSetup[];
+
+  getSetup(): SnifferConfigSetup[];
+  update(
+    existingId: string,
+    newConfig: SnifferConfig,
+    isStarted: boolean
+  ): void;
+  addSniffer(snifferConfig: SnifferConfig): void;
+  removeSniffer(port: number): void;
+  setIsStarted(snifferId: string, isStarted: boolean): void;
+}
